Move credentials exceptions to central location
