+++ python/Ganga/Core/exceptions.py
@@ -270,6 +270,34 @@ class BulkOperationRepositoryError(RepositoryError):
     def getOriginalJobError(self, id):
         return self.details.get(id)
 
+
+class CredentialsError(GangaException):
+    """
+    Base class for credential-related errors
+    """
+    pass
+
+
+class CredentialRenewalError(CredentialsError):
+    """
+    There was some problem with renewing a credential
+    """
+    pass
+
+
+class InvalidCredentialError(CredentialsError):
+    """
+    The credential is invalid for some reason
+    """
+    pass
+
+
+class ExpiredCredentialError(InvalidCredentialError):
+    """
+    The credential has expired
+    """
+    pass
+
 #
 #
 # $Log: not supported by cvs2svn $
